change to a simple module of existing aambackendservice
