+++ application/aam-backend-service/src/main/kotlin/com/aamdigital/aambackendservice/thirdpartyauthentication/repository/AuthenticationSessionRepository.kt
@@ -1,0 +1,10 @@
+package com.aamdigital.aambackendservice.thirdpartyauthentication.repository
+
+import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.data.repository.PagingAndSortingRepository
+import java.util.*
+
+interface AuthenticationSessionRepository : JpaRepository<AuthenticationSessionEntity, Long>,
+    PagingAndSortingRepository<AuthenticationSessionEntity, Long> {
+    fun findByExternalIdentifier(externalIdentifier: String): Optional<AuthenticationSessionEntity>
+}
