--- application/aam-integration/settings.gradle.kts
@@ -1,1 +1,0 @@
-rootProject.name = "aam-integration"
