--- application/aam-integration/src/main/kotlin/com/aamdigital/aamintegration/Application.kt
@@ -1,30 +1,0 @@
-package com.aamdigital.aamintegration
-
-import org.springframework.boot.autoconfigure.SpringBootApplication
-import org.springframework.boot.context.properties.ConfigurationPropertiesScan
-import org.springframework.boot.runApplication
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories
-import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.RestController
-import java.util.*
-
-@SpringBootApplication
-@ConfigurationPropertiesScan
-@EnableJpaRepositories
-@EnableMethodSecurity
-class Application
-
-fun main(args: Array<String>) {
-    TimeZone.setDefault(TimeZone.getTimeZone("UTC"))
-    runApplication<Application>(*args)
-}
-
-@RestController
-class HelloFriendController {
-    @GetMapping("/")
-    @Suppress("FunctionOnlyReturningConstant")
-    fun helloFriend(): String {
-        return "Hello, Friend. This is aam-integration."
-    }
-}
