--- application/aam-integration/src/main/kotlin/com/aamdigital/aamintegration/authentication/core/VerifySessionUseCase.kt
@@ -1,16 +1,0 @@
-package com.aamdigital.aamintegration.authentication.core
-
-import com.aamdigital.aamintegration.domain.DomainUseCase
-import com.aamdigital.aamintegration.domain.UseCaseData
-import com.aamdigital.aamintegration.domain.UseCaseRequest
-
-data class VerifySessionUseCaseRequest(
-    val sessionId: String,
-    val sessionToken: String,
-) : UseCaseRequest
-
-data class VerifySessionUseCaseData(
-    val userId: String,
-) : UseCaseData
-
-abstract class VerifySessionUseCase : DomainUseCase<VerifySessionUseCaseRequest, VerifySessionUseCaseData>()
