--- application/aam-integration/src/main/kotlin/com/aamdigital/aamintegration/authentication/repository/AuthenticationSessionRepository.kt
@@ -1,10 +1,0 @@
-package com.aamdigital.aamintegration.authentication.repository
-
-import org.springframework.data.jpa.repository.JpaRepository
-import org.springframework.data.repository.PagingAndSortingRepository
-import java.util.*
-
-interface AuthenticationSessionRepository : JpaRepository<AuthenticationSessionEntity, Long>,
-    PagingAndSortingRepository<AuthenticationSessionEntity, Long> {
-    fun findByExternalIdentifier(externalIdentifier: String): Optional<AuthenticationSessionEntity>
-}
