--- application/aam-integration/src/main/kotlin/com/aamdigital/aamintegration/error/HttpErrorDto.kt
@@ -1,6 +1,0 @@
-package com.aamdigital.aamintegration.error
-
-open class HttpErrorDto(
-    val errorCode: String,
-    val errorMessage: String,
-)
