--- application/aam-integration/src/main/kotlin/com/aamdigital/aamintegration/rest/ObjectMapperConfiguration.kt
@@ -1,35 +1,0 @@
-package com.aamdigital.aamintegration.rest
-
-import com.fasterxml.jackson.databind.DeserializationFeature
-import com.fasterxml.jackson.databind.ObjectMapper
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.context.annotation.Primary
-import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder
-import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter
-
-
-@Configuration
-class ObjectMapperConfiguration {
-
-    @Bean
-    @Primary
-    fun objectMapper(): ObjectMapper {
-        val mapper = Jackson2ObjectMapperBuilder()
-        mapper.featuresToEnable(
-            DeserializationFeature.ACCEPT_EMPTY_STRING_AS_NULL_OBJECT,
-            DeserializationFeature.READ_UNKNOWN_ENUM_VALUES_USING_DEFAULT_VALUE
-        )
-        return mapper.build()
-    }
-
-    @Bean
-    fun mappingJackson2HttpMessageConverter(): MappingJackson2HttpMessageConverter {
-        val builder = Jackson2ObjectMapperBuilder()
-        builder.featuresToEnable(
-            DeserializationFeature.ACCEPT_EMPTY_STRING_AS_NULL_OBJECT,
-            DeserializationFeature.READ_UNKNOWN_ENUM_VALUES_USING_DEFAULT_VALUE
-        )
-        return MappingJackson2HttpMessageConverter(builder.build())
-    }
-}
