--- application/aam-integration/src/main/resources/application.yaml
@@ -1,92 +1,0 @@
-spring:
-  application:
-    name: aam-integration
-  main:
-    banner-mode: off
-  jackson:
-    deserialization:
-      accept-empty-string-as-null-object: true
-      read-unknown-enum-values-using-default-value: true
-  datasource:
-    driver-class-name: org.postgresql.Driver
-    username: admin
-    password: docker
-  jpa:
-    generate-ddl: true
-    hibernate:
-      ddl-auto: update
-  threads:
-    virtual:
-      enabled: true
-
-server:
-  servlet:
-    context-path: /
-
-management:
-  endpoint:
-    health:
-      probes:
-        enabled: true
-  endpoints:
-    web:
-      exposure:
-        include:
-          - info
-          - health
-  tracing:
-    sampling:
-      probability: 1
-
-features:
-  third-party-authentication:
-    enabled: false
-
-logging:
-  logback:
-    rollingpolicy:
-      max-file-size: 10MB
-      total-size-cap: 10MB
-
----
-# <second part of the config, just active with profile "local-development">
-
-spring:
-  config:
-    activate:
-      on-profile: local-development
-  security:
-    oauth2:
-      resourceserver:
-        jwt:
-          issuer-uri: http://localhost:8888/realms/third-party-integration
-  datasource:
-    driver-class-name: org.postgresql.Driver
-    url: jdbc:postgresql://localhost:5403/aam-integration
-    username: admin
-    password: docker
-
-server:
-  error:
-    include-message: always
-    include-binding-errors: always
-  port: 9001
-
-logging:
-  level:
-    #    org.springframework.amqp.rabbit: warn
-    #    org.springframework.web: debug
-    #    org.springframework.http: debug
-    #    org.springframework.security: debug
-    com.aamdigital.aamintegration: trace
-
-features:
-  third-party-authentication:
-    enabled: true
-
-aam-keycloak-config:
-  realm: master
-  client-id: aam-integration
-  client-secret: 1234
-  server-url: http://localhost:8888
-  application-url: aam.localhost
